+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support `file:///` protocol for index URL
+
+### Fixed
+- Custom indexes were not usable in interactive mode
+- Temporary downloads clean up on error
 
 ## [0.7.0]
 ### Added